configuration was divided into two files
