+++ dz4/src/main/java/ru/otus/dz4/config/ApplicationSettings.java
@@ -11,7 +11,6 @@ import org.springframework.context.support.ReloadableResourceBundleMessageSource
 public class ApplicationSettings {
 
     private String[] csvFile;
-    private String messageBundlePath;
 
     public void setCsvFile(String csvFile) {
         String[] strings = csvFile.split(",");
@@ -22,20 +21,4 @@ public class ApplicationSettings {
         return csvFile;
     }
 
-    public void setMessageBundlePath(String messageBundlePath) {
-        this.messageBundlePath = messageBundlePath;
-    }
-
-    public String getMessageBundlePath() {
-        return messageBundlePath;
-    }
-
-    @Bean
-    MessageSource messageSource() {
-        ReloadableResourceBundleMessageSource ms = new ReloadableResourceBundleMessageSource();
-//        ms.setBasename("/i18n/bundle");
-        ms.setBasename(getMessageBundlePath());
-        ms.setDefaultEncoding("UTF-8");
-        return ms;
-    }
 }
